+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Note: Breaking changes between versions are indicated by "💥".
 ## Unreleased
 
 - [Bugfix] Fix "Invalid Request" error during SAML authentication (thanks @Ant1x!).
-- [Feature] Add `images build --pull` option to update base images.
+- [Feature] Add `make pull-base-images` command to update base images.
 - [Improvement] Annotate types all over the Tutor code base.
 - [Bugfix] Fix parsing of YAML CLI arguments that include equal "=" signs.
 - [Bugfix] Fix minor edge case in `long_to_base64` utility function.